% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{is_spatial}
\alias{is_spatial}
\title{Is an object a spatial object?}
\usage{
is_spatial(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{terra}, \code{sf}, \code{raster}, or \code{sp} spatial object.
}
\description{
Is an object a spatial object?
}
